+++ casual/casual-api/src/test/groovy/se/laz/casual/network/ProtocolVersionTest.groovy
@@ -1,30 +1,54 @@
 /*
- * Copyright (c) 2022, The casual project. All rights reserved.
+ * Copyright (c) 2022 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
 package se.laz.casual.network
 
+
 import spock.lang.Specification
 
 class ProtocolVersionTest extends Specification
 {
-   def 'unmarshalling'()
+   def "unmarshalling"()
    {
-      when:
-      ProtocolVersion protocolVersion = ProtocolVersion.unmarshall(version)
-      then:
-      protocolVersion.name() == versionString
-      where:
-      version | versionString
-      1000L   | 'VERSION_1_0'
-      1001L   | 'VERSION_1_1'
-      1002L   | 'VERSION_1_2'
+      expect:
+      for( ProtocolVersion p : ProtocolVersion.values(  ) )
+      {
+         assert ProtocolVersion.unmarshall( p.getVersion(  ) ) == p
+         assert ProtocolVersion.unmarshall( p.getVersionAsString(  ) ) == p
+      }
    }
 
-   def 'unsupported version'()
+   def "unsupported version long"()
    {
+      when:
+      ProtocolVersion.unmarshall( value )
 
+      then:
+      thrown ProtocolVersionException
+
+      where:
+      value << [
+              0L,
+              -1L
+      ]
+   }
+
+   def "Unsupported version string"()
+   {
+      when:
+      ProtocolVersion.unmarshall( value )
+
+      then:
+      thrown ProtocolVersionException
+
+      where:
+      value << [
+              null,
+              "",
+              "12"
+      ]
    }
 
 }