+++ casual/casual-inbound/src/main/java/se/laz/casual/jca/inflow/CasualMessageListenerImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2024, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -29,11 +29,16 @@ import se.laz.casual.jca.inbound.handler.service.ServiceHandler;
 import se.laz.casual.jca.inbound.handler.service.ServiceHandlerFactory;
 import se.laz.casual.jca.inbound.handler.service.ServiceHandlerNotFoundException;
 import se.laz.casual.jca.inflow.work.CasualServiceCallWork;
+import se.laz.casual.network.InboundDeactivatedContext;
+import se.laz.casual.network.InboundTopologyUpdateContext;
+import se.laz.casual.network.ProtocolMatcher;
+import se.laz.casual.network.ProtocolVersion;
 import se.laz.casual.network.protocol.messages.CasualNWMessageImpl;
 import se.laz.casual.network.protocol.messages.domain.CasualDomainConnectReplyMessage;
 import se.laz.casual.network.protocol.messages.domain.CasualDomainConnectRequestMessage;
 import se.laz.casual.network.protocol.messages.domain.CasualDomainDiscoveryReplyMessage;
 import se.laz.casual.network.protocol.messages.domain.CasualDomainDiscoveryRequestMessage;
+import se.laz.casual.network.protocol.messages.domain.DomainDisconnectReplyMessage;
 import se.laz.casual.network.protocol.messages.domain.Service;
 import se.laz.casual.network.protocol.messages.service.CasualServiceCallRequestMessage;
 import se.laz.casual.network.protocol.messages.transaction.CasualTransactionResourceCommitReplyMessage;
@@ -65,25 +70,45 @@ import java.util.logging.Logger;
 public class CasualMessageListenerImpl implements CasualMessageListener
 {
     private static Logger log = Logger.getLogger(CasualMessageListenerImpl.class.getName());
-    private static final Long CASUAL_PROTOCOL_VERSION = 1000L;
-
     @Override
     public void domainConnectRequest(CasualNWMessage<CasualDomainConnectRequestMessage> message, Channel channel)
     {
         log.finest(() -> "domainConnectRequest(). " + PrettyPrinter.format(message.getCorrelationId(), message.getMessage().getExecution()) + message );
+        log.info(()-> "domainConnectRequest(). client" + channel + " asking for protocol version(s)" + message.getMessage().getProtocols());
+        log.info(()-> "domainConnectRequest(). supported protocols: " + ProtocolVersion.supportedVersions());
+        Long matchedProtocolVersion = ProtocolMatcher.match(message.getMessage().getProtocols());
+        log.info(() -> "domainConnectRequest(). matched protocol version: " + ProtocolVersion.unmarshall(matchedProtocolVersion));
 
+        if(matchedProtocolVersion >= ProtocolVersion.VERSION_1_1.getVersion())
+        {
+            // should be notified when RA is deactivated
+            InboundDeactivatedContext.add(channel);
+        }
+        if(matchedProtocolVersion >= ProtocolVersion.VERSION_1_2.getVersion())
+        {
+            // should be notified whenever a domain connects
+            // we send to all previously connected clients and add the new one after as not to send
+            // the notification to the new client connecting
+            InboundTopologyUpdateContext.sendTopologyUpdate(message.getMessage().getExecution());
+            InboundTopologyUpdateContext.add(channel);
+        }
         String domainName = ConfigurationService.getConfiguration( ConfigurationOptions.CASUAL_DOMAIN_NAME );
         UUID domainId = ConfigurationService.getConfiguration( ConfigurationOptions.CASUAL_DOMAIN_ID ).getId();
         CasualDomainConnectReplyMessage reply = CasualDomainConnectReplyMessage.createBuilder()
                                                                                .withDomainId( domainId )
                                                                                .withDomainName( domainName )
                                                                                .withExecution( message.getMessage().getExecution() )
-                                                                               .withProtocolVersion(CASUAL_PROTOCOL_VERSION)
+                                                                               .withProtocolVersion(matchedProtocolVersion)
                                                                                .build();
         CasualNWMessage<CasualDomainConnectReplyMessage> replyMessage = CasualNWMessageImpl.of( message.getCorrelationId(), reply );
         channel.writeAndFlush(replyMessage);
     }
 
+    @Override
+    public void domainDisconnectReply(CasualNWMessage<DomainDisconnectReplyMessage> message)
+    {
+        log.finest(() -> "domainDisconnectReply(). " + PrettyPrinter.format(message.getCorrelationId(), message.getMessage().getExecution()) + message );
+    }
 
     @Override
     public void domainDiscoveryRequest(CasualNWMessage<CasualDomainDiscoveryRequestMessage> message, Channel channel)