+++ casual/casual-network/src/main/java/se/laz/casual/network/inbound/ExceptionHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2018, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -9,6 +9,8 @@ package se.laz.casual.network.inbound;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
+import se.laz.casual.network.InboundDeactivatedContext;
+import se.laz.casual.network.InboundTopologyUpdateContext;
 
 import java.util.logging.Logger;
 
@@ -25,6 +27,8 @@ public final class ExceptionHandler extends ChannelInboundHandlerAdapter
     public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause)
     {
         log.warning(() -> "casual inbound exception caught: " + cause + " closing channel");
+        InboundDeactivatedContext.remove(ctx.channel());
+        InboundTopologyUpdateContext.remove(ctx.channel());
         ctx.close();
     }
 }