+++ casual/casual-network/src/test/groovy/se/laz/casual/network/inbound/ExceptionHandlerTest.groovy
@@ -1,12 +1,14 @@
 /*
- * Copyright (c) 2017 - 2018, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
 
 package se.laz.casual.network.inbound
 
+import io.netty.channel.Channel
 import io.netty.channel.ChannelHandlerContext
+import io.netty.channel.ChannelId
 import spock.lang.Specification
 
 class ExceptionHandlerTest extends Specification
@@ -14,7 +16,11 @@ class ExceptionHandlerTest extends Specification
     def 'should close context'()
     {
         setup:
-        def ctx = Mock(ChannelHandlerContext)
+        def ctx = Mock(ChannelHandlerContext){
+           channel() >> Mock(Channel){
+              id() >> Mock(ChannelId)
+           }
+        }
         def handler = ExceptionHandler.of()
         when:
         handler.exceptionCaught(ctx, new RuntimeException())