+++ casual/casual-network/src/test/groovy/se/laz/casual/network/utils/FakeListener.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2018, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2025, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -7,18 +7,19 @@
 package se.laz.casual.network.utils
 
 import io.netty.channel.Channel
+import jakarta.resource.spi.XATerminator
+import jakarta.resource.spi.endpoint.MessageEndpoint
+import jakarta.resource.spi.work.WorkManager
 import se.laz.casual.api.network.protocol.messages.CasualNWMessage
 import se.laz.casual.jca.inflow.CasualMessageListener
 import se.laz.casual.network.protocol.messages.domain.CasualDomainConnectRequestMessage
 import se.laz.casual.network.protocol.messages.domain.CasualDomainDiscoveryRequestMessage
+import se.laz.casual.network.protocol.messages.domain.DomainDisconnectReplyMessage
 import se.laz.casual.network.protocol.messages.service.CasualServiceCallRequestMessage
 import se.laz.casual.network.protocol.messages.transaction.CasualTransactionResourceCommitRequestMessage
 import se.laz.casual.network.protocol.messages.transaction.CasualTransactionResourcePrepareRequestMessage
 import se.laz.casual.network.protocol.messages.transaction.CasualTransactionResourceRollbackRequestMessage
 
-import jakarta.resource.spi.XATerminator
-import jakarta.resource.spi.endpoint.MessageEndpoint
-import jakarta.resource.spi.work.WorkManager
 import java.lang.reflect.Method
 
 class FakeListener implements MessageEndpoint, CasualMessageListener
@@ -43,7 +44,13 @@ class FakeListener implements MessageEndpoint, CasualMessageListener
 
     }
 
-    @Override
+   @Override
+   void domainDisconnectReply(CasualNWMessage<DomainDisconnectReplyMessage> message)
+   {
+
+   }
+
+   @Override
     void domainDiscoveryRequest(CasualNWMessage<CasualDomainDiscoveryRequestMessage> message, Channel channel) {
 
     }